FZF InterOp (#1)
* Prefer junegunn/fzf for selection if available

This change brings in better UX for selecting targets. FZF has become
the staple fuzzy-find adapter in the Vim world. Leveraging this should
make the targets feel nicer to work with.

Writing this out too, I realize using FZF is easier for me programming
as well. Vim's `popup` returns with the index of the selected item in
the list — which requires one extra hop to work with said item — but
with FZF's sink, you get the actual value in the sink. Less code.

* Update documentation to reference FZF and Dispatch