Select all from targets within scope (#2)
* Select from all targets within scope

It was too good to be true that the selector always chose the
package:package target when filtering. I realized that this was a bug
because I didn't include the spread glyph™ in the Bazel query. This
change fixes that.

* Fixup UI for non-FZF experience

BoB uses the Vim 9 popup to select the target being built. This change
just updates its look and feel to look something closer to FZF. This
change also prefers the low-level `popup_create` to the `popup_menu`
abstraction.