bazel run (#3)
* No need to manually update mtime

* More garbage lines

* Add Run entrypoint

* Refactor Build to pass in FuncRef

* Decouple "build" from building process

There needs to be a better name for this. I just cannot think of this
now. Ideally, it's the one process which takes care of Build and Run. I
leaned towards Execute, but that means something else within Vim and
could be confusing. It should work for now, but I should find something
better.

p.s: There's no more permanent solution than a temporary solution 😌

* Wire up Run