Late guards
Because of the way Vim loads, the plugin we depend on may not be loaded
before ours is loaded. I don't know if this is because of the vim-plug
plugin manager, or actually how Vim loads at startup. But for some
reason, vim-dispatch is not loaded at the time the plugin is being
loaded.

To solve for this, I move the check for the dependency close to where
the dependency is needed. This way, Bob can still do its job until he
absolutely gets to where the capability to do his job is lacking.